feat(smaller modules): breaking up  entities into smaller modules
* Implement new modules with existing business logic
* Added a registry to avoid circulair imports
