+++ tests/cassettes/test_oktalib_apps_test_get_saml_application_metadata.json
@@ -25,7 +25,7 @@
             "keep-alive"
           ],
           "Content-Length": [
-            "2313"
+            "2351"
           ],
           "Content-Type": [
             "text/xml;charset=ISO-8859-1"
